Slight reorg to clean some things up and make things clearer.
