+++ afar/__init__.py
@@ -1,5 +1,6 @@
+from ._core import get, run  # noqa
 from ._version import get_versions
-from .core import get, later, locally, remotely, run  # noqa
+from ._where import later, locally, remotely  # noqa
 
 __version__ = get_versions()["version"]
 del get_versions
