+++ afar/_exceptions.py
@@ -1,0 +1,2 @@
+class AfarException(Exception):
+    """Used to explicitly indicate where and how to execute the code of a context"""
